feat(App) : 기능 통합 정리
